+++ test/unit/fuzz-tools.test.js
@@ -102,6 +102,7 @@ import * as idcritV99 from '../../lib/idcrit-v99.js';
 import * as cardioV101 from '../../lib/cardio-v101.js';
 import * as cardioV102 from '../../lib/cardio-v102.js';
 import * as cvriskV103 from '../../lib/cvrisk-v103.js';
+import * as cardioV104 from '../../lib/cardio-v104.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -148,6 +149,7 @@ const MODULES = {
   'cardio-v101.js': cardioV101,
   'cardio-v102.js': cardioV102,
   'cvrisk-v103.js': cvriskV103,
+  'cardio-v104.js': cardioV104,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];